chore: split main into modules
